feat(fitchef): add coach insight runtime wrapper
